Moved logic for creative creation to higher level.
